Refactor build scripts: remove unused https_curl function and update libz-sys build configuration
