+++ patches/libz-sys/build.rs
@@ -3,10 +3,12 @@ fn main() {
 
     if target == "armv6k-nintendo-3ds" {
         // Use the pre-built portlib
-        println!("cargo:rustc-link-search=native=/opt/devkitpro/portlibs/3ds/lib");
+        let devkitpro = std::env::var("DEVKITPRO").expect("DEVKITPRO not set");
+        let portlibs = format!("{devkitpro}/portlibs/3ds");
+        println!("cargo:rustc-link-search=native={portlibs}/lib");
         println!("cargo:rustc-link-lib=static=z");
-        println!("cargo:root=/opt/devkitpro/portlibs/3ds");
-        println!("cargo:include=/opt/devkitpro/portlibs/3ds/include");
+        println!("cargo:root={portlibs}");
+        println!("cargo:include={portlibs}/include");
     } else {
         // Use real libz-sys behaviour
         println!("cargo:rustc-link-lib=z");
